third_party/blink/web_tests/fast/dom/Window/orphaned-frame-access-expected.txt

This is a testharness.js-based test.
[FAIL] Indexed property access on detached Window
  assert_equals: Indexed property should still be accessible. expected (string) "123" but got (undefined) undefined
Harness: the test ran to completion.
